feat: move tests to proper directory with English localization
- Move AgentScannerTests.cs to test/Aevatar.GAgents.AI.Abstractions.Test/
- Remove temporary test files from root directory
- Translate all Chinese content to English (comments, strings, console output)
- Fix AgentIndexInfo and AgentDescriptionAttribute structure mismatch
- Adjust L1Description length to meet 100-150 character requirement
- All 6 unit tests now pass successfully
- Establish professional English testing framework for Agent infrastructure
